Started working on celebration crud operations.
